refactor: refact providers folder
